Add a popover similar to the Tap table for the Source and Destination columns in the Top UI
